Use tracing in verify method too
